requirements.txt
3to2==1.1.1
appdirs==1.4.3
applib==1.2
apturl==0.5.2
args==0.1.0
backports.weakref==1.0rc1
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
Brlapi==0.6.4
certifi==2017.7.27.1
chardet==3.0.4
checkbox-support==0.22
click==6.7
clint==0.5.1
cmdln==2.0.0
CodeIntel==2.0.0rc2
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
decorator==4.1.2
deep-belief-network==1.0.1
defer==1.0.6
dlib==19.7.99
EasyProcess==0.2.3
entrypoints==0.2.3
enum34==1.1.6
esprima==4.0.0.dev12
feedparser==5.1.3
floyd-cli==0.10.17
gdbn==0.1
gnumpy==0.2
guacamole==0.9.2
html5lib==0.9999999
httplib2==0.9.1
idna==2.6
Inflector==2.0.11
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
jedi==0.11.0
Jinja2==2.8
joblib==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
labelImg==1.5.1
language-selector==0.1
Lasagne==0.1
libclang-py3==3.9.0
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.9
MarkupSafe==0.23
marshmallow==2.14.0
matplotlib==2.1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
nolearn==0.6.0
nose==1.3.7
notebook==5.1.0
numpy==1.13.3
oauthlib==1.0.3
object-detection==0.1
onboard==1.2.0
padme==1.1.1
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.1.2
plainbox==0.25
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==5.4.3
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
pytz==2017.2
PyVirtualDisplay==0.2.1
pyxdg==0.25
pyzmq==16.0.2
qtconsole==4.3.1
raven==6.2.1
reportlab==3.3.0
requests==2.18.4
requests-toolbelt==0.8.0
scikit-learn==0.18.1
scipy==0.18.1
screen-resolution-extra==0.0.0
selenium==3.8.1
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.11.0
sklearn==0.0
stevedore==1.27.1
system-service==0.3
tabulate==0.8.1
tensorflow==1.4.0
tensorflow-gpu==1.2.1
tensorflow-tensorboard==0.4.0rc2
terminado==0.6
testpath==0.3.1
tflearn==0.3.2
Theano==0.9.0
tornado==4.5.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.3
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zope.cachedescriptors==4.3.1
